Improve tool error handling
It's normal for RAID members not to appear in diskutil output. We don't want to treat this as an error, so we'll return None instead of raising DeviceNotFoundError.

Similarly, it's normal for smartctl to return non-zero statuses when a health check fails, there are errors in the device logs, etc. We now properly parse the return value and return None if it indicates that the command itself failed, as opposed to the disk's status being bad.